refactor: user/statistics/domain/dto 패키지를 user/statistics/dto로 이동
