+++ src/main/java/com/upstage/devup/user/statistics/controller/UserStatController.java
@@ -1,8 +1,8 @@
 package com.upstage.devup.user.statistics.controller;
 
 import com.upstage.devup.auth.config.AuthenticatedUser;
-import com.upstage.devup.user.statistics.domain.dto.UserSolvedQuestionDto;
-import com.upstage.devup.user.statistics.domain.dto.WrongNoteSummaryDto;
+import com.upstage.devup.user.statistics.dto.UserSolvedQuestionDto;
+import com.upstage.devup.user.statistics.dto.WrongNoteSummaryDto;
 import com.upstage.devup.user.statistics.service.UserAnswerStatService;
 import com.upstage.devup.user.statistics.service.UserWrongAnswerReadService;
 import lombok.RequiredArgsConstructor;
