+++ src/test/java/com/upstage/devup/user/statistics/controller/UserStatControllerTest.java
@@ -3,8 +3,8 @@ package com.upstage.devup.user.statistics.controller;
 import com.upstage.devup.auth.config.AuthenticatedUser;
 import com.upstage.devup.auth.config.SecurityConfig;
 import com.upstage.devup.auth.config.jwt.JwtTokenProvider;
-import com.upstage.devup.user.statistics.domain.dto.UserSolvedQuestionDto;
-import com.upstage.devup.user.statistics.domain.dto.WrongNoteSummaryDto;
+import com.upstage.devup.user.statistics.dto.UserSolvedQuestionDto;
+import com.upstage.devup.user.statistics.dto.WrongNoteSummaryDto;
 import com.upstage.devup.user.statistics.service.UserAnswerStatService;
 import com.upstage.devup.user.statistics.service.UserWrongAnswerReadService;
 import org.junit.jupiter.api.DisplayName;
